% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenLFChains}
\alias{eigenLFChains}
\title{Post-process MCMC samples}
\usage{
eigenLFChains(
  splineS,
  splineT,
  mod,
  numeig,
  iter,
  burnin,
  nchains,
  s,
  t,
  alpha
)
}
\arguments{
\item{splineS}{basis matrix in the longitudinal direction}

\item{splineT}{basis matrix in the functional direction}

\item{mod}{an object returned from mcmcWeakChains}

\item{numeig}{number of eigenfunctions to infer}

\item{iter}{number of total iterations in the original mcmc}

\item{burnin}{number of these iterations to use as burnin}

\item{nchains}{number of chains}

\item{s}{complete data longitudinal times}

\item{t}{complete data functional times}
}
\value{
Posterior marginal eigenfunctions, eigenvalues, mean, and 
their associated uncertainty.
}
\description{
This function will post-process posterior draws from an object returned from
mcmcWeakChains.
}
\examples{
See the example in Root/Example
}
