% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_marginal_func}
\alias{get_marginal_func}
\title{Get marginal functional covariance matrix}
\usage{
get_marginal_func(cov, ns, nt)
}
\arguments{
\item{cov}{An ns*nt by ns*nt dimensional covariance matrix}

\item{ns}{Number of longitudinal points}

\item{nt}{Number of functional points}
}
\value{
An nt x nt marginal functional covariance matrix
}
\description{
Integrate over the longitudinal dimension to obtain an nt x nt marginal
covariance matrix
}
\examples{
See Root/Simulation
}
