% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_marginal_long}
\alias{get_marginal_long}
\title{Get marginal longitudinal covariance matrix}
\usage{
get_marginal_long(cov, ns, nt)
}
\arguments{
\item{cov}{An ns*nt by ns*nt dimensional covariance matrix}

\item{ns}{Number of longitudinal points}

\item{nt}{Number of functional points}
}
\value{
An ns x ns marginal longitudinal covariance
}
\description{
Integrate over the functional dimension to obtain an ns x ns marginal
covariance matrix
}
\examples{
See Root/Simulation
}
