% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik}
\alias{loglik}
\title{Calculate log-likelihood}
\usage{
loglik(y, X, Bs, Bt, missing, Theta, Varphi, iter, burnin)
}
\arguments{
\item{y}{A list of of length n containing responses}

\item{X}{An n x p design matrix}

\item{Bs}{Basis matrix for longitudinal direction}

\item{Bt}{Basis matrix for functional direction}

\item{missing}{A list of length n containing missing indices for each
response}

\item{Theta}{Posterior draws of Theta, arranges in a cube}

\item{Varphi}{Posterior draws of Varphi, vector format}

\item{iter}{Number of total samples}

\item{burnin}{Number of samples to use as burnin}
}
\value{
A Matrix of size (iter - burnin) x number of observed time points
over all subjects containing log-likelihood values
}
\description{
Calculates log-likelihood after obtaining draws from the joint posterior
distribution. Use after running run_mcmc.
}
