% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run Markov-Chain Monte-Carlo sampling algorithm}
\usage{
run_mcmc(y, missing, X, splineS, splineT, q1, q2, iter, thin, burnin, nchains)
}
\arguments{
\item{y}{A list of of length n containing responses}

\item{missing}{A list of length n containing missing indices for each
response}

\item{X}{An n by p design matrix}

\item{splineS}{Basis matrix for longitudinal direction}

\item{splineT}{Basis matrix for functional direction}

\item{q1}{Number of latent factors for longitudinal direction}

\item{q2}{Number of latent factors for functional direction}

\item{iter}{Number of posterior samples to keep}

\item{thin}{Keep every thin samples}

\item{burnin}{Number of burnin samples to discard for posterior inference}

\item{nchains}{How many chains to run}
}
\value{
A list of samples for each parameter. This list can be used as input
to eigenLFChains for further post-processing
}
\description{
This function will generate samples from the posterior distribution.
}
\examples{
See the example in Root/Example
}
